railway

+++ src/main/java/com/app/api/service/UserService.java
@@ -8,7 +8,7 @@ public interface UserService {
 	public abstract void insert(Users user);
 	public abstract void update(Users user);
 	//public abstract void delete(Integer userId);
-	public abstract Users findById(Integer userId);
+	public abstract Users findById(Long userId);
 	public abstract Users findByUsername(String username);
 	public abstract Collection<Users> findAll();
 }
